refactor handleAddNode
